ai-server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paho-mqtt==1.6.1
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
jinja2==3.1.4
